Split some functions to new files.
