+++ graphics/model_puma560.py
@@ -1,0 +1,3 @@
+def import_puma_560():
+    # TODO
+    pass
